dataoutput: the user can decide to have data written to an ascii file or for julia to plot and write the plots to a png file. :outformat => "png" or "ascii" in user_inputs.jl. Default if "png"
